+++ .prettierrc
@@ -1,0 +1,6 @@
+{
+  "printWidth": 100,
+  "tabWidth": 2,
+  "trailingComma": "all",
+  "semi": true
+}
