+++ README.md
@@ -1,7 +1,7 @@
-# WNRS - We're Not Really Strangers
+# WNRS
 
-Play We're Not Really Strangers on your browser [here](https://munjoonteo.github.io/wnrs/)! Contains all the cards from the standard deck with a few choice additions.
+Play WNRS on your browser [here](https://munjoonteo.github.io/wnrs/)! Contains all the cards from the standard deck with a few choice additions.
 
-Created in collaboration with [@ilyues](https://github.com/ilyues).  
+Created in collaboration with [@ilyues](https://github.com/ilyues).
 
 Written with React.
