+++ public/robots.txt
@@ -1,3 +1,4 @@
-# https://www.robotstxt.org/robotstxt.html
-User-agent: *
-Disallow:
+# https://www.robotstxt.org/robotstxt.html	
+User-agent: *	
+Disallow:	
+
