+++ src/assets/levels.ts
@@ -1,0 +1,175 @@
+export const levelOne = [
+  "What was your first impression of me?",
+  "What do you think is the hardest part of what I do for a living",
+  "What reality show do you think I'm most likely to binge watch? explain.",
+  "Do I seem like a cat or dog person?",
+  "What do my shoes tell you about me?",
+  "Do I seem like more of a creative or analytical type? explain.",
+  "How many speeding tickets do you think I've gotten in my life?",
+  "Do I seem like someone who would get a name tattooed on myself?",
+  "What do you think I'm most likely to splurge on?",
+  "Finish the sentence: just by looking at you I'd think ____.",
+  "What character would I play in a movie?",
+  "Do I remind you of anyone?",
+  "Who do you think my celebrity crush is?",
+  "Do you think I've ever checked an ex's phone for evidence?",
+  "What's the first thing you noticed about me?",
+  "Do I seem like a coffee or tea person? Sweetened or unsweetened?",
+  "What do you think my go to karaoke song is?",
+  "If you were to buy me a present, knowing nothing about me other than what I look like, what would it be?",
+  "What does my phone wallpaper tell you about me?",
+  "Do I look kind? explain.",
+  "Do you think I fall in love easily? why or why not?",
+  "As a child, what do you think I wanted to be?",
+  "Do you think I'm usually early, on time, or late to events? explain.",
+  "What subject do you think I thrived in at school? Did I fail any?",
+  "What compliment do you think I hear the most?",
+  "Do I seem like a morning person or a night owl? why?",
+  "What fast food restaurant do you think I'm most likely to drive through? What's my order?",
+  "What is my body language telling you right now?",
+  "What does my instagram tell you about me?",
+  "Do you think plants thrive or die in my care? explain.",
+  "On a scale of 1-10, how messy do you think my car is? 1 being clean 10 complete disaster",
+  "Do you think I was popular in school? explain.",
+  "Do you think I've ever been fired from a job? If so, what for?",
+  "Do you think I intimidate others? why or why not?",
+  "How likely am I to go camping? how high maintenance is my set up?",
+  "What about me is most strange or unfamiliar to you?",
+  "If myspace were still a thing, what would my profile song be?",
+  "What about me intrigues you?",
+  "What does my style tell you about me?",
+  "Do you think I've ever had my heart broken?",
+  "Do you think I've been in love?",
+  "What is the worst assumption someone has made about you?",
+];
+
+export const levelTwo = [
+  "What is your defining characteristic?",
+  "What's the most pain you've ever been in that wasn't physical?",
+  "How are you, really?",
+  "What is the last thing your mom has lied to you about?",
+  "What's your father's name and one thing about him?",
+  "What have I tolerated from people in the past that I no longer have space for?",
+  "Something that brought a smile to my face this week was ___.",
+  "What are you feeling a lot of lately?",
+  "What do you need right now but aren't communicating?",
+  "What are you currently working through that others don't see?",
+  "What was the biggest turning point in your dating life?",
+  "What is your heart telling you?",
+  "What is your most toxic trait you can admit to?",
+  "Who in your life deserves the biggest thank you? Let them know if you can.",
+  "What are you overthinking right now?",
+  "What did you last relationship teach you about yourself?",
+  "What belief about yourself no longer serves you?",
+  "What do you need to forgive yourself for?",
+  "What's the hardest lesson you've learned in love?",
+  "You hurt me, but thank you for ___.",
+  "Who in your life can you be your most vulnerable with?",
+  "Which one of your parent's personality traits do you want to keep/let go of?",
+  "Dear younger self: ____.",
+  "Think of someone you admire. Why did this person come to mind?",
+  "Are you lying to yourself about anything?",
+  "What's your favorite part of yourself that isn't physical?",
+  "What's the hardest part about dating you?",
+  "I feel loved when ___.",
+  "What is something you've learned about yourself this week?",
+  "Self love is ___.",
+  "What was the lesson in your most recent painful experience?",
+  "I know I'm in love when ___.",
+  "What's your favorite song lyric you can think of off the top of your head?",
+  "How did you get over your first love?",
+  "What's the best lesson an ex has ever taught you?",
+  "Is there anyone who's changed your life but doesn't know it?",
+  "What do you need right now, more than anything?",
+  "When was the last time you felt lucky to be you?",
+  "What do you need to accept today?",
+  "When was the last time you felt most yourself?",
+  "What's your mother's name and the most valuable lesson she's taught you thus far?",
+  "What are you most grateful for in this current moment?",
+  "What are you still trying to prove to yourself?",
+  "What have you been taking for granted lately?",
+  "Is there a feeling you miss?",
+  "What would your younger self not believe about your life today?",
+  "If you could have it your way: Who would you be with? Where would you be? And what would you be doing?",
+  "What do you need to hear right now?",
+  "What title would you give this chapter in your life?",
+  "What dating advice would you give your younger self?",
+  "What lesson should you have learned by now?",
+  "When was the last time you were able to use your pain to help someone else?",
+  "What would you never want to change about yourself?",
+  "What question did this year leave you with?",
+  "What about yourself is hardest to admit?",
+  "What question are you trying to answer most in your life right now?",
+  "Finish the sentences: strangers would describe me as _____. Only I know that I am ______",
+  "When you're asked how are you, how often do you answer truthfully?",
+  "Has a stranger ever changed your life?",
+  "What is the most unexplainable thing that's ever happened to you?",
+  "Describe your perfect day!",
+  "Have you changed your mind about anything recently?",
+  "What's been the best compliment a stranger has ever given you?",
+  "What is a compliment you wish you received more frequently?",
+  "What do you crave more of?",
+  "What has been your earliest recollection of happiness?",
+  "What are you more afraid of, failure or success? And why?",
+  "How would you describe the feeling of being in love in one word?",
+  "When was the last time you surprised yourself?",
+  "What's been your happiest memory this past year?",
+  "Do you think the image you have of yourself matches the image people see you as?",
+  "If you have, when was the moment you realized you weren't invincible?",
+  "What lesson took you the longest to unlearn?",
+  "What part of your life works? What part of your life hurts?",
+  "What is your 1st love's name and the reason you fell in love with him/her?",
+  "What's your mother's name? And the most beautiful thing about her?",
+  "If you could get to know someone in your life on a deeper level, who would it be and why?",
+  "Are you missing anyone right now? Do you think they are missing you too?",
+  "What is a dream you've let go of?",
+  "What is the last thing you lied to your mother about?",
+  "Have you ever told someone I love you but didn't mean it? if so, why?",
+  "How can you become a better person?",
+];
+
+export const levelThree = [
+  "What can I help you with?",
+  "Why do you think we met?",
+  "What about me surprised you?",
+  "In a word, how would you describe our conversation?",
+  "What do you think my strength is?",
+  "What can I teach you?",
+  "Based on what you've learned about me, do you have any Netflix recommendations?",
+  "Who do I remind you of?",
+  "What can you be kinder towards yourself for today?",
+  "What's one small thing you can do to be a better person in 2019?",
+  "What can we create together?",
+  "How would you describe me to a stranger?",
+  "What about me most surprised you?",
+  "What would be the perfect gift for me?",
+  "What do you admire most about me?",
+  "Based on what you learned about me, what book would you recommend I read?",
+  "What about me is hardest for you to understand?",
+  "What am I most qualified to give advice about?",
+  "What is one thing you think I can do that would dramatically improve my life?",
+  "If we were in a band, what would our name be?",
+  "Based on what you learned about me, does my social media accurately reflect who I am? Why or why not?",
+  "What do I need to hear right now?",
+  "Based on what you know about me, do you have any netflix recommendations?",
+  "What do you think my superpower is?",
+  "What is a lesson you will take away from our conversation?",
+  "When in this game did you feel most connected to me?",
+  "How does one earn your vulnerability? Have I earned it? How can I earn more?",
+  "What do you think our most important similarity is?",
+  "In one word, how would you describe our conversation?",
+  "What answer of mine made you light up?",
+  "When this game is over, what will you remember about me?",
+  "What question were you most afraid to answer?",
+  "What parts of yourself do you see in me?",
+  "How do our personalities complement each other?",
+  "What do you think my weakness is?",
+  "What would make you feel closer to me?",
+  "What do you think I fear the most?",
+  "What has this conversation taught you about yourself?",
+  "What do you think I should know about myself that perhaps I'm unaware of?",
+  "What do you think my defining characterisitic is?",
+  "Do you believe everyone has a calling? If so, do you think I've found mine?",
+  "What do you recommend I let go of, if anything?",
+  "In one word, describe how you feel right now",
+];
