+++ src/components/card/Card.css
@@ -5,7 +5,6 @@
   justify-content: center;
   text-transform: uppercase;
   font-family: "Biryani", sans-serif;
-  
 }
 
 .big-card {
